requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.7
django-allauth==64.0.0
django-crispy-forms==2.3
django-import-export==4.1.1
django-admin-generator==2.6.0
django-crispy-forms==2.3
django-phonenumber-field==8.0.0
django-utils2==3.0.2
phonenumbers==8.13.40
python-utils==3.8.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
psycopg2-binary==2.9.5
thinkific==0.1.1
gunicorn==23.0.0
django-browser-reload==1.21.0